bnc#402315 - Make the main-menu fit in small monitors
If the default position of the main-menu window would make it not fit in the monitor's
area, we would move the menu's window to be aligned with the *other* side of the main-menu button.
This would cause it to be almost completely outside the screen on small monitors.
Now we just clamp the position of the window to the monitor's bounds.

Signed-off-by: Federico Mena Quintero <federico@novell.com>